only crete dir when they don't exist

+++ python_app/file_system/file_system_main.py
@@ -346,8 +346,9 @@ class SectionInfoStructure:
             
             sectionFolderName = pathToTopSection.split("/")[-1]
             mainTemplateFile = os.path.join(os.getenv("BOOKS_TEMPLATES_PATH"),"main_template.tex")
-            _waitDummy = os.makedirs(os.path.join(pathToTopSection,"_out"))
-            _waitDummy = shutil.copy(mainTemplateFile, os.path.join(pathToTopSection, sectionFolderName + "_main.tex"))
+            if not os.path.exists(os.path.join(pathToTopSection,"_out")):
+                _waitDummy = os.makedirs(os.path.join(pathToTopSection,"_out"))
+                _waitDummy = shutil.copy(mainTemplateFile, os.path.join(pathToTopSection, sectionFolderName + "_main.tex"))
 
     def _getPathToSectionsFolder():
         pathToSectionFolder = _u.Settings.readProperty(_u.Settings.PubProp.currBookPath_ID)
